==> notebooks/03_baseline_linear_regression.ipynb <==
# Imports

import numpy as np                  # Numerical operations
import pandas as pd                 # Data handling and CSV I/O
import joblib                       # Save/load Python objects (models, arrays)

# Machine Learning Models
from sklearn.linear_model import LinearRegression  # Baseline regression model

# Evaluation Metrics
from sklearn.metrics import mean_absolute_error, mean_squared_error, r2_score  # Model performance metrics


# Load Preprocessed Data

# Load Preprocessed Data
X_train = joblib.load(r"D:\Python\SmartSignalAI\data\X_train.pkl")
y_train = joblib.load(r"D:\Python\SmartSignalAI\data\y_train.pkl")

X_val   = joblib.load(r"D:\Python\SmartSignalAI\data\X_val.pkl")
y_val   = joblib.load(r"D:\Python\SmartSignalAI\data\y_val.pkl")

X_test  = joblib.load(r"D:\Python\SmartSignalAI\data\X_test.pkl")
y_test  = joblib.load(r"D:\Python\SmartSignalAI\data\y_test.pkl")

print("Preprocessed data loaded successfully!")
print("Train shape:", X_train.shape, y_train.shape)
print("Validation shape:", X_val.shape, y_val.shape)
print("Test shape:", X_test.shape, y_test.shape)

# Train Linear Regression

lr_model = LinearRegression() # lr_model - This is the variable name

# Train Model
lr_model.fit(X_train, y_train)   # Fit on training data only

# Validation Evaluation

# predict on the validation set to check model generalization
# - MAE: average error in dBm  
# - RMSE: penalizes large errors, important for network coverage  
# - R²: variance explained by the model

# Validation Model

y_val_pred = lr_model.predict(X_val) 

mae_val  = mean_absolute_error(y_val, y_val_pred) 
rmse_val = np.sqrt(mean_squared_error(y_val, y_val_pred))
r2_val   = r2_score(y_val, y_val_pred)

print("Linear Regression - Validation Metrics:")
print("MAE:", round(mae_val, 3))
print("RMSE:", round(rmse_val, 3))
print("R²:", round(r2_val, 3))

# Test Evaluation

# Test Model

y_test_pred = lr_model.predict(X_test)

mae_test  = mean_absolute_error(y_test, y_test_pred)
rmse_test = np.sqrt(mean_squared_error(y_test, y_test_pred))
r2_test   = r2_score(y_test, y_test_pred)

print("Linear Regression - Test Metrics")
print("MAE:", round(mae_test, 3))
print("RMSE:", round(rmse_test, 3))
print("R²:", round(r2_test, 3))

# Test set is used **once**, after validation, for unbiased final metrics.
# Ensures no leakage from train/validation.

# Save Linear Regression Model

# Save the trained model -Stored using joblib in `data/linear_model.pkl`.

joblib.dump(lr_model, r"D:\Python\SmartSignalAI\data\linear_model.pkl")
print("Linear Regression Model Saved Successfully!")
